requirements.txt
# Core dependencies
pandas>=2.0.0
openai>=1.0.0
python-dotenv>=0.19.0
tqdm>=4.64.0
openpyxl>=3.1.0

# PDF processing
PyPDF2>=3.0.0

# Optional: for improved Excel handling
xlsxwriter>=3.0.0
